Move lsSymbolKind to lsp.h
